Update checkpointing reference and add basic accumulation example
- Added src/basic-accumulation-example.py

- Updated reference in src/basic-checkpointing.py

- Modified reference in src/feature-hashing-ctr-example.py

+++ utils/cmd/cleanup.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 for file in README*; do
-  if [ "$file" != "README.org" ]; then
+  if [ "$file" != "README.org" ] && [ "$file" != "README.trash.org" ]; then
     rm "$file"
   fi
 done
